KYLIN-1924 Region server metrics: replace int type for long type for scanned row count

+++ core-cube/src/main/java/org/apache/kylin/cube/inmemcubing/ConcurrentDiskStore.java
@@ -154,7 +154,7 @@ public class ConcurrentDiskStore implements IGTStore, Closeable {
         final DataInputStream din;
         long fileLen;
         long readOffset;
-        int count;
+        long count;
 
         Reader(long startOffset) throws IOException {
             this.fileLen = diskFile.length();
@@ -265,7 +265,7 @@ public class ConcurrentDiskStore implements IGTStore, Closeable {
         }
 
         @Override
-        public int getScannedRowCount() {
+        public long getScannedRowCount() {
             return count;
         }
