+++ core-cube/src/main/java/org/apache/kylin/gridtable/GTFilterScanner.java
@@ -63,7 +63,7 @@ public class GTFilterScanner implements IGTScanner {
     }
 
     @Override
-    public int getScannedRowCount() {
+    public long getScannedRowCount() {
         return inputScanner.getScannedRowCount();
     }
